🚚 Rename search-options to grouped-network
