+++ apps/web/app/layout.tsx
@@ -9,7 +9,7 @@ import { SpeedInsights } from "@vercel/speed-insights/next";
 
 // utils
 import localFont from "next/font/local";
-import { getGroupedNetworkChains } from "~/lib/search-options";
+import { getGroupedNetworkChains } from "~/lib/grouped-network-chains";
 
 // types
 import type { Metadata } from "next";
